refactor to state and ui controller
